+++ static/js/server_events.js
@@ -122,8 +122,11 @@ function dispatch_normal_event(event) {
                 subs.mark_subscribed(sub.name, sub);
             });
         } else if (event.op === 'peer_add') {
+            // TODO: remove email shim here and fix called functions
+            //       to use user_ids
+            var person = people.get_person_from_user_id(event.user_id);
+            var email = person.email;
             _.each(event.subscriptions, function (sub) {
-                var email = event.user_email;
                 stream_data.add_subscriber(sub, email);
                 $(document).trigger('peer_subscribe.zulip',
                                     {stream_name: sub, user_email: email});